working on adding new user
