+++ forlocalsclient/src/Routes/index.js
@@ -5,7 +5,6 @@ import BusinessDetails from '../Views/BusinessDetails';
 import BusinessProfile from '../Views/BusinessProfile';
 import Home from '../Views/Home';
 import LocalProfile from '../Views/LocalProfile';
-import Login from '../Views/Login';
 import Search from '../Views/Search';
 import BusinessProfileDetails from '../Views/BusinessProfileDetails';
 import AddBusiness from '../Views/AddBusiness';
@@ -17,11 +16,10 @@ export default function Routing({ local }){
                 <Route path="/"  element={<Home local={local}/>} />
                 <Route path="/search"  element={<Search local={local}/>} />
                 <Route path="/businessDetails/:id"  element={<BusinessDetails local={local}/>} />
-                <Route path="/login"  element={<Login local={local}/>} />
                 <Route path="/businessProfile/:firebaseKey" element={<BusinessProfile local={local}/>} />
-                <Route path="/localProfile/:firebaseKey"  element={local ? <LocalProfile local={local}/> : <Login /> } />
+                <Route path="/localProfile/:firebaseKey"  element={local ? <LocalProfile local={local}/> : <Home /> } />
                 <Route path="/businessManager/:id" element={<BusinessProfileDetails local={local} />} />
-                <Route path="registerBusiness/:firebaseKey" element={local ? <AddBusiness local={local} /> : <Login />} />           
+                <Route path="registerBusiness/:firebaseKey" element={local ? <AddBusiness local={local} /> : <Home />} />           
             </Routes>
         </>
     )
